requirements.txt
numpy
Pillow
pyyaml

openvino
